refactor(chain): decouple finalized endpoint profiles
